+++ test/factories/ForeignChainContractFactory.test.ts
@@ -8,31 +8,34 @@ import {
   GenericForeignChainContractProps,
   IGenericForeignChainContract,
 } from '../../src/contracts/generic/IGenericForeignChainContract';
-import { ForeignChainsIds, NonEvmChainsIds } from '../../src/types/ChainsIds';
+import { ChainsIds, ForeignChainsIds, NonEvmChainsIds } from '../../src/types/ChainsIds';
 import settings from '../../src/config/settings';
 
 describe('ForeignChainContractFactory', () => {
   describe('#create', () => {
     describe('Given a foreign chainId for an evm blockchain', async () => {
-      ForeignChainsIds.filter((x) => !NonEvmChainsIds.includes(x)).forEach((chainId) => {
-        it(`should return a ForeignChainContract instance for ${chainId}`, async () => {
-          const blockchain = BlockchainFactory.create({ chainId, settings });
-          expect(!!blockchain).to.eql(true);
-          expect(blockchain.chainId).to.eql(chainId);
+      ForeignChainsIds.filter((x) => !NonEvmChainsIds.includes(x))
+        .filter((x) => x !== ChainsIds.ARBITRUM) // TODO enable when we move to goerli
+        .forEach((chainId) => {
+          it(`should return a ForeignChainContract instance for ${chainId}`, async () => {
+            const blockchain = BlockchainFactory.create({ chainId, settings });
+            expect(blockchain).to.not.be.undefined;
+            expect(blockchain.chainId).to.eql(chainId);
 
-          const foreignChainContract = ForeignChainContractFactory.create(<BaseChainContractProps>{
-            blockchain: blockchain,
-            settings,
+            const foreignChainContract = ForeignChainContractFactory.create(<BaseChainContractProps>{
+              blockchain: blockchain,
+              settings,
+            });
+
+            expect(foreignChainContract).to.not.be.empty;
+
+            await foreignChainContract.resolveContract();
+            expect((<ForeignChainContract>foreignChainContract).address()).to.not.empty;
+
+            const address = foreignChainContract.address();
+            expect(typeof address).to.eql('string');
           });
-
-          await foreignChainContract.resolveContract();
-          expect(!!foreignChainContract).to.eql(true);
-          expect(!!(<ForeignChainContract>foreignChainContract).address()).to.eql(true);
-
-          const address = foreignChainContract.address();
-          expect(typeof address).to.eql('string');
         });
-      });
     });
 
     describe('Given a foreign chainId for a non-evm blockchain', async () => {